supporting caanoo, line doublers, refactoring
git-svn-id: file:///home/notaz/opt/svn/PicoDrive@893 be3aeb3a-fb24-0410-a615-afba39da0efa

+++ platform/gp2x/soc.c
@@ -39,7 +39,7 @@ gp2x_soc_t soc_detect(void)
 	int memdev;
 	int i;
 
-	if (ret != -2)
+	if ((int)ret != -2)
 		/* already detected */
 		return ret;
